[ADD] filtering navigation added - parsing bugs fixed

+++ app/src/main/java/com/light/finder/data/source/remote/dto/ProductBrowsingDto.kt
@@ -7,7 +7,6 @@ data class ProductBrowsingListDto(
 )
 
 data class ProductBrowsingDto(
-
     @SerializedName("product_country") val product_country: String,
     @SerializedName("product_promoted") val product_promoted: Int,
     @SerializedName("product_discount_proc") val product_discount_proc: Int,
@@ -18,7 +17,7 @@ data class ProductBrowsingDto(
     @SerializedName("product_cct_code") val product_cct_code: Int,
     @SerializedName("product_dimming_code") val product_dimming_code: Int,
     @SerializedName("product_finish") val product_finish: String,
-    @SerializedName("product_image_no_exten") val product_image_no_exten: Int,
+    @SerializedName("product_image_no_exten") val product_image_no_exten: Long,
     @SerializedName("product_name") val product_name: String,
     @SerializedName("product_prio") val product_prio: Int,
     @SerializedName("product_category_name") val product_category_name: String,
@@ -26,9 +25,9 @@ data class ProductBrowsingDto(
     @SerializedName("product_wattage_replaced") val product_wattage_replaced: Int,
     @SerializedName("product_price_pack") val product_price_pack: Double,
     @SerializedName("product_price_sku") val product_price_sku: Double,
-    @SerializedName("product_price_lamp") val product_price_lamp: Int,
-    @SerializedName("product_SAPid_10NC") val product_SAPid_10NC: Int,
-    @SerializedName("product_SAPid_12NC") val product_SAPid_12NC: Int,
+    @SerializedName("product_price_lamp") val product_price_lamp: Double,
+    @SerializedName("product_SAPid_10NC") val product_SAPid_10NC: Long,
+    @SerializedName("product_SAPid_12NC") val product_SAPid_12NC: Long,
     @SerializedName("product_qty_lampsku") val product_qty_lampsku: Int,
     @SerializedName("product_qty_skucase") val product_qty_skucase: Int,
     @SerializedName("product_qty_lampscase") val product_qty_lampscase: Int,
